pyproject.toml
[tool.poetry]
name = "crawler-helper"
version = "0.1.0"
description = ""
authors = ["umatbro <umatbroo@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
requests = "^2.22"
beautifulsoup4 = "^4.8"
django-dotenv = "^1.4"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8"
celery = "^4.3"
sendgrid = "^6.1"

[tool.poetry.dev-dependencies]
pre-commit = "^1.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
